+++ src/main/java/com/imt/Api_monstres/controller/dto/input/MonsterHttpDto.java
@@ -13,10 +13,10 @@ import lombok.RequiredArgsConstructor;
 public class MonsterHttpDto {
     private final String playerId;
     private final Elementary element; 
-    private final Double hp;
-    private final Double atk;
-    private final Double def;
-    private final Double vit;
+    private final Integer hp;
+    private final Integer atk;
+    private final Integer def;
+    private final Integer vit;
     private final List<SkillHttpDto> skills;
     private final Rank rank;
